api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlglot
pydantic
openai>=1.0.0
langgraph